# backend/src/main/java/fr/epita/assistants/ping/api/request/MoveRequest.java
package fr.epita.assistants.ping.api.request;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class MoveRequest {
    private String src;
    private String dst;
}