# backend/src/main/java/fr/epita/assistants/ping/api/request/PathRequest.java
package fr.epita.assistants.ping.api.request;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class PathRequest {
    private String relativePath;
}